feat: add jacoco, update dependencies, and configure firebase
Adds the JaCoCo plugin for test coverage and enables it in the debug build type. The `google()` repository is added to the buildscript.

Dependencies are updated, including Firebase BOM, Compose BOM, and Lifecycle KTX. The MockK dependency setup is also revised.

Additionally, the `google-services.json` is updated with a new Firebase project configuration.