feat: add workflow to automatically create issues from TODO comments
This commit introduces a new GitHub Actions workflow named "Run TODO to Issue".

This workflow is triggered on push events or can be manually dispatched. It checks out the code, uses the "alstr/todo-to-issue-action" to find TODO comments, creates corresponding GitHub issues, and inserts the issue URLs back into the code.

If changes are made (i.e., issue URLs are inserted), the workflow commits these changes with the message "Automatically added GitHub issue links to TODOs" and pushes them to the main branch.